move apply_chunk() to a separate file
