+++ tools/state-viewer/src/lib.rs
@@ -1,6 +1,7 @@
 #![doc = include_str!("../README.md")]
 
 mod apply_chain_range;
+mod apply_chunk;
 pub mod cli;
 mod commands;
 mod epoch_info;
